+++ src/utils/index.js
@@ -17,4 +17,31 @@ function shuffleAnswers(answers) {
   return answersCopy
 }
 
-export { shuffleAnswers }
+function calculateNewScore({
+  answeredQuestionsCount,
+  correctAnswersCount,
+  QUESTIONS_LENGTH
+}) {
+  if (
+    QUESTIONS_LENGTH <= 0 ||
+    correctAnswersCount < 0 ||
+    answeredQuestionsCount < 0
+  )
+    return null
+
+  const currentScore =
+    answeredQuestionsCount < 1
+      ? 0
+      : (correctAnswersCount * 100) / answeredQuestionsCount
+
+  const maxScore =
+    ((QUESTIONS_LENGTH - (answeredQuestionsCount - correctAnswersCount)) *
+      100) /
+    QUESTIONS_LENGTH
+
+  const minScore = (correctAnswersCount * 100) / QUESTIONS_LENGTH
+
+  return { currentScore, maxScore, minScore }
+}
+
+export { shuffleAnswers, calculateNewScore }